Update ftplugins for C and C++ files
The commenting functions for the c and cpp ftplugins are changed so that
commented lines can be uncommented regardless of their initial indents.
Additionally, both plugins' undo functions are expanded to reverse the
effects of the plugins properly; originally, remapping errors occur when
switching to a different file type that defines a mapping with the same
left-hand side ("lhs") as that of a mapping defined in the initial file
type.