Add summaries of the Vim user manual
Summaries documenting my learning journey of Vim via reading the Vim
user manual are added. They can aid in skimming particular chapters of
the manual in the future. Also added is a table of contents so that all
the summaries can be glanced. The tags file is updated with the tags
introduced by these files.