Update Makefile template
The template file for Makefiles is revised to improve readability.
Additionally, debug and release targets and their corresponding flags
are added. Finally, the line for the target's name (where the cursor is
positioned when creating a Makefile in Vim) is fixed near the top of the
file so that future amendments to the template file do not necessitate
changing the .vimrc file.