Change variable flavor and revise rules
Previously, variables were defined as "recursively expanded" variables
with operator=, which means their values can change depending on when
they are used. Now, they are defined as "simply expanded" variables (via
operator:=) so that they behave more like the variables in most
programming languages: their values are exactly the same as when they
were defined.

Additionally, the executable targets "debug" and "release" are prefixed
with the binary directory so that "make" actually knows where to find
them (to determine whether a rebuild is necessary).

Finally, the binary and object directories are specified directly for
the rules that need them as prerequisites. This change allows dependency
relationships to be identified more easily.