Revise template file for Makefiles
Much of the template file for Makefiles is rewritten. The major changes
are adding a way to rebuild the default goal if a header file is updated
and removing redefinitions of implicit variables. Other changes include
renaming variables, defining additional variables for other directories,
specifying new phony targets, and general revision for readability.