Add mappings and a command
A command that behaves like ":retab" but has an inverse relationship
with 'expandtab' is added. This command makes reversing ":retab" faster
and more convenient. Also, Command-line mappings that emulate the
behaviors of some GNU-Readline keyboard shortcuts are defined. This
change makes editing on the terminal and Vim's Command line more
consistent.