Add more GNU-Readline shortcuts
Accidentally deleting characters with CTRL-W or all of them with CTRL-U
on the Command line is annoying: Vim does not provide a way to undo the
deletion. Therefore, this commit overwrites the Command-line command
CTRL-Y and overloads Command-line CTRL-U and CTRL-W to produce behaviors
similar to the corresponding keys in GNU Readline. That is, CTRL-U and
CTRL-W behave as normal, but whatever they delete is stored in a local
buffer and can be retrieved with CTRL-Y.