Specify <SID> for expression mappings
Vim9 script allows expression mappings to be evaluated in the contexts
of the scripts in which they are defined. This capability means that the
special "<SID>" string is not needed for such mappings to refer to
script-local functions. However, the important phrase is "in which they
are defined"; sometimes, these mappings will not be defined in the same
script (for example, creating a session script via ":mksession" and
sourcing it). When such a scenario occurs, the expression mappings no
longer have access to their specified script-local items, and errors
such as "E117: Unknown function" will surface. Therefore, this commit
avoids this problem by retaining <SID> for script-local functions in
expression mappings.