Finish reformatting GNU 'make' documentation
The process of reformatting the plain-text documentation for GNU 'make'
into a Vim help file is complete; the document is reformatted according
to the guidelines given in "help-writing". Specific edits include
changing single-quote strings into double-quote ones, rounding code
blocks to the next indent level, replacing in-text references with the
appropriate tags, and formatting text to a text-width of 78.

The tags file is updated accordingly.