Terminate code blocks properly
According to "help-writing", code blocks are terminated when a '<'
character is the first non-blank on a line following such blocks or a
line starting in column 1 succeeds them. Some code blocks in make.txt
are followed by conventionally formatted paragraphs: their first lines
are indented. Since these lines do not begin at the first column, they
are considered to be a part of the preceding code blocks. Thus, this
commit addresses this formatting error by inserting a '<' at the start
of such lines.