Improve Update_last_change()
Update_last_change() is made more robust: a format string may be
supplied to use a particular style for the new date; the first 20 lines
are scanned for a "Last change:" string; and 'filetype' is no longer
limited to "help". Additionally, this function is now invoked whenever
any file is written, not just text files.