Add options, key bindings, and aliases
The EXTENDED_GLOB zsh option is enabled to exploit zsh's extended
globbing operators. Additionally, key bindings for undoing and redoing
changes on the command-line buffer are added. A key binding for listing
the shell's jobs is also added. Finally, aliases for programs that
fulfill specific task categories are added so that such programs can be
replaced without needing to memorize new names; if "vimx" exists, "vim"
is set as its alias to ensure Vim has access to the X-server clipboard
by default.