Permit remapping and support Wayland clipboard
The mappings defined for accessing the system clipboard are elevated to
regular mappings: their right-hand sides may be analyzed for further
mappings. This change is made so that the behavior of their right-hand
sides can be adjusted for specific window systems, which is the second
change of this commit: if Wayland is used, quoteplus ("+) register puts,
yanks, and deletes are redefined to use the "wl-clipboard" utility to
access the clipboard on the Wayland window system.