Include Control 'hjkl' keys for selecting panes
Key bindings for 'C-h', 'C-j', 'C-k', and 'C-l' are added. They can be
used identically to their plain-key counterparts (without the Control
key modifier) to select panes in the corresponding directions. This
change is made to follow Vim's pattern of binding both a character and
its Control-key version to the same window-command behavior.

To avoid precluding the prefix key from being sent to the terminal (it
might be used to invoke a different tmux command rather than
"send-prefix"), the backslash key is used to send the prefix key.