Add system-specific options
An option specific to macOS systems is added. This option specifies the
default shell command to which text copies made in tmux's copy mode
should be piped. In this case, the command is "pbcopy", which is a
builtin command of macOS that provides access to the system clipboard
from the command line.