Fix Command-line buffer problem with unprintables
Previously, Backward_kill_pre() and Put_cmdline_buffer() returned the
contents of the Command-line and buffer holding the most recent text
deleted, respectively, as part of their strings for mapping expressions.
This implementation for a Command-line buffer, however, had a limitation
in which unprintable characters, particularly <C-M> and <Nul>, in such
contents would interfere with the mapping expressions' execution.

This commit removes this limitation by eliminating the opportunity for
unprintable characters to be passed to mapping expressions.
Backward_kill_pre() and Backward_kill_post() now use the CMDLINE
dictionary for communication rather than arguments and parameters.
Additionally, Put_cmdline_buffer()'s return value includes a call to an
accessor which retrieves the contents of the Command-line buffer instead
of containing the buffer's contents directly. Although inadvertent
unprintable characters are not being passed to mapping expressions, the
Command-line buffer can still get and put these characters.