Remove the <special> argument from mappings
The <special> argument allows mappings to use the <> notation to denote
special keys even if the "<" flag appears in 'cpoptions'. However, this
effect currently is not useful because 'cpoptions' would contain "<"
only if it was set manually. Since the vimrc file does not mutate
'cpoptions', <special> can be removed safely.