Add mappings to edit files in argument list
This commit defines mappings that jump to the first, second, third, or
fourth file in the argument list. This change enables quick switching
between files in a set without needing to remember file names or buffer
numbers. Additionally, this commit adds mappings that replace such files
in the argument list with the current file. The keys used to edit or
replace these files are ALT-H, ALT-T, ALT-N, and ALT-S, which are the
home keys for the right hand on the Dvorak keyboard layout.