Remove trick to toggle app launcher
The implementation for toggling the application launcher is fragile and
imperfect: keys defined to exit the launcher in the launcher's
configuration file need to be mirrored in this implementation, and
selecting an item in the launcher does not exit "launcher" mode, causing
the next input of the key binding to start the launcher to be ignored.
Removing this toggle capability completely simplifies accessing the
application launcher.