Toggle syntax highlighting correctly
The functions treesitter.stop() and treesitter.start() from Neovim do
not work as expected as they only apply to one buffer and generate an
error if tree-sitter cannot find a language for the buffer. Conversely,
the plugin commands ":TSEnable" and ":TSDisable" from
nvim-treesitter/nvim-treesitter are correct as they execute silently and
apply to all buffers, like ":syntax off" and ":syntax on".