Reformat mappings for consistency
Mappings with a Lua function on the right-hand side are indented to be
consistent with other Lua-function mappings in the same file. Similarly,
the key sequence "<Leader>" is made all lowercase to match its uses in
the diagnostic keymaps and plugin specifications. Also, newlines are
added among these mappings to make them more readable.