Update Nvim config according to Nvim v0.10.0
Some mappings defined in the custom configuration for Nvim are defined
by default in Nvim v0.10.0; thus, they are removed. Also,
vim.diagnostic.is_disabled() and vim.diagnostic.disable() are
deprecated, and the signature for vim.diagnostic.enable() is changed.
The mapping to toggle diagnostics is updated according to these changes.