Use a shell script to set Sway variables in kanshi
To simplify the configuration for kanshi and avoid working with the
intricacies of its exec command, such as the '$' character not being
able to be passed in the format required by the set command of Sway.

Also, update the path to the zshrc file for the scratchpad terminal
according to the change made in Commit f6986a3 (Rename dotfiles with
"dot-" prefix, 2024-07-25).